[#109] impl: StringFromUTF8Lossy
